+++ public/index.html
@@ -4,10 +4,9 @@
 <head>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <meta name="theme-color" content="#000000" />
   <meta name="description" content="Web site created using create-react-app" />
-  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-  <title>React App</title>
+  <link rel="icon" href="images/logo_outline.png" />
+  <title>Care3</title>
 </head>
 
 <body>
