+++ src/Components/ChatRoomDate/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { Date, DateP } from "./style.js";
+
+function ChatRoomDate() {
+  return (
+    <Date>
+      <DateP>Apr 1</DateP>
+    </Date>
+  );
+}
+
+export default ChatRoomDate;
