+++ src/Components/ChatRoomDate/style.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const Date = styled.div`
+  text-align : center;
+`;
+
+export const DateP = styled.p`
+  font-size : 11px;
+  color     : #aaaaaa;
+`;
