+++ src/Components/ChatRoomSupitle/index.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Head, HeadPFirst, HeadPSecond } from "./style.js";
+
+function ChatRoomSupitle() {
+  return (
+    <Head>
+      <HeadPFirst>Starter Care Plan created + daily actions added</HeadPFirst>
+      <HeadPSecond>12:10p</HeadPSecond>
+    </Head>
+  );
+}
+
+export default ChatRoomSupitle;
